move the functions in solvers/Utilities.xpp into utilities.xpp
