+++ problems/NLMCProblems.cpp
@@ -1,6 +1,6 @@
 #include "mfem.hpp"
 #include "NLMCProblems.hpp"
-#include "../solvers/Utilities.hpp"
+#include "../utilities.hpp"
