+++ utilities.hpp
@@ -1,4 +1,8 @@
 #include "mfem.hpp"
+#ifdef MFEM_USE_STRUMPACK
+#include <StrumpackOptions.hpp>
+#include <mfem/linalg/strumpack.hpp>
+#endif
 
 #ifndef UTILITY_FUNCTIONS
 #define UTILITY_FUNCTIONS
@@ -18,5 +22,26 @@ mfem::HypreParMatrix * GenerateProjector(HYPRE_BigInt * offsets, HYPRE_BigInt * 
 
 HYPRE_BigInt * offsetsFromLocalSizes(int n, MPI_Comm comm = MPI_COMM_WORLD);
 
+mfem::HypreParMatrix * NonZeroRowMap(const mfem::HypreParMatrix& A);
+
+mfem::HypreParMatrix * NonZeroColMap(const mfem::HypreParMatrix& A);
+
+
+class DirectSolver : public mfem::Solver
+{
+private:
+   mfem::Solver* solver;
+#ifdef MFEM_USE_STRUMPACK
+    mfem::STRUMPACKRowLocMatrix* Astrumpack;
+#endif
+
+public:
+   DirectSolver();
+   DirectSolver(const mfem::Operator& op);
+   virtual ~DirectSolver();
+
+   void SetOperator(const mfem::Operator& op) override;
+   void Mult(const mfem::Vector &b, mfem::Vector &x) const override;  
+};
 
 #endif
